# [P4Testgen] Create abstract execution state with utility functions - move copy-in/out to control/parser blocks.