… backends/p4tools/common/CMakeLists.txt
… backends/p4tools/common/core/abstract_execution_state.h
… backends/p4tools/common/core/target.cpp
… backends/p4tools/common/core/target.h
… backends/p4tools/common/lib/variables.h
… backends/p4tools/modules/testgen/core/program_info.cpp
… backends/p4tools/modules/testgen/core/small_step/cmd_stepper.cpp
… backends/p4tools/modules/testgen/core/small_step/cmd_stepper.h
… backends/p4tools/modules/testgen/core/small_step/expr_stepper.cpp
… backends/p4tools/modules/testgen/core/small_step/extern_stepper.cpp
… backends/p4tools/modules/testgen/core/small_step/table_stepper.cpp
… backends/p4tools/modules/testgen/lib/collect_coverable_nodes.cpp
… backends/p4tools/modules/testgen/lib/execution_state.cpp
… backends/p4tools/modules/testgen/lib/execution_state.h
… backends/p4tools/modules/testgen/targets/bmv2/cmd_stepper.cpp
… backends/p4tools/modules/testgen/targets/bmv2/expr_stepper.cpp
… backends/p4tools/modules/testgen/targets/bmv2/program_info.cpp
… backends/p4tools/modules/testgen/targets/bmv2/table_stepper.cpp
… backends/p4tools/modules/testgen/targets/ebpf/cmd_stepper.cpp
… backends/p4tools/modules/testgen/targets/ebpf/program_info.cpp
… backends/p4tools/modules/testgen/targets/pna/dpdk/cmd_stepper.cpp
… backends/p4tools/modules/testgen/targets/pna/dpdk/program_info.cpp
… backends/p4tools/modules/testgen/targets/pna/shared_table_stepper.cpp
… backends/p4tools/modules/testgen/test/lib/taint.cpp
… ir/irutils.cpp
… testdata/p4_16_samples/gauntlet_lookahead_expansion-bmv2.p4
… testdata/p4_16_samples/gauntlet_lookahead_expansion.p4
… testdata/p4_16_samples_outputs/gauntlet_lookahead_expansion-bmv2-first.p4
… testdata/p4_16_samples_outputs/gauntlet_lookahead_expansion-bmv2-frontend.p4
… testdata/p4_16_samples_outputs/gauntlet_lookahead_expansion-bmv2-midend.p4