Other P4Tools modules can also benefit from execution state, but Testgen's execution state was too specialized.  This PR cleans up the execution state class and introduces an abstract, general execution state. Coupled with that, we introduce several utilities that are beneficial for performance and simplification of core functions. Among others, we simplify the way we handle copy-in/out and struct assignments. 
There were also other small improvements:

- Added some more table utilities.
- Reduce some unnecessary execution state copies.
- Do not offload parser state resolution to the IR::PathExpression visitor. Instead, resolve select expressions and parser states directly.



Performance for generating 10000 tests for `pins_middleblock.p4`.
Before:
```
Executed in  495.09 secs    fish           external
   usr time  483.90 secs  409.00 micros  483.90 secs
   sys time    8.98 secs  252.00 micros    8.98 secs
```

After:
```
Executed in  431.90 secs    fish           external
   usr time  422.09 secs  519.00 micros  422.09 secs
   sys time    7.86 secs  320.00 micros    7.86 secs

```
A ~13% reduction.